Moved doControlRequest function into actions
