+++ client/app/scripts/components/node-details/node-details-table-row.js
@@ -102,19 +102,6 @@ function renderValues(node, columns = [], columnStyles = [], timestamp = null, t
   });
 }
 
-/**
- * Table row children may react to onClick events but the row
- * itself does detect a click by looking at onMouseUp. To stop
- * the bubbling of clicks on child elements we need to dismiss
- * the onMouseUp event.
- */
-export const dismissRowClickProps = {
-  onMouseUp: (ev) => {
-    ev.preventDefault();
-    ev.stopPropagation();
-  }
-};
-
 export default class NodeDetailsTableRow extends React.Component {
   constructor(props, context) {
     super(props, context);
@@ -186,7 +173,7 @@ export default class NodeDetailsTableRow extends React.Component {
         onMouseLeave={this.onMouseLeave}
         className={className}>
         <td className="node-details-table-node-label truncate" style={firstColumnStyle}>
-          {this.props.renderIdCell(Object.assign(node, {nodeId, topologyId}))}
+          {this.props.renderIdCell(Object.assign(node, { nodeId, topologyId }))}
         </td>
         {values}
       </tr>
