+++ client/app/scripts/utils/dom-utils.js
@@ -14,3 +14,16 @@ export function encodeIdAttribute(id) {
 export function decodeIdAttribute(id) {
   return id.replace(/__u(\d+)__/gm, (m, d) => String.fromCharCode(d));
 }
+
+/**
+ * Table row children may react to onClick events but the row
+ * itself does detect a click by looking at onMouseUp. To stop
+ * the bubbling of clicks on child elements we need to dismiss
+ * the onMouseUp event.
+ */
+export const dismissRowClickProps = {
+  onMouseUp: (ev) => {
+    ev.preventDefault();
+    ev.stopPropagation();
+  }
+};
